ocr-service/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
numpy<2.0
paddleocr==2.9.1
paddlepaddle==3.0.0
Pillow==11.2.1
python-multipart==0.0.20
